+++ plugins/transformer/jsonflattener/resources/plugin/plugin.go
@@ -1,0 +1,9 @@
+package plugin
+
+// Don't move this file to a different package, it's used by Go releaser to embed the version in the binary.
+var (
+	Name    = "jsonflattener"
+	Kind    = "transformer"
+	Team    = "cloudquery"
+	Version = "development"
+)